refactor(lynx): restore loading prop in ActionButton package, simplify snippet to thin wrapper
- Remove loading from Omit in ActionButton props, add loading variant to CSS recipe call
- When loading=true, render children directly (not wrapped in <text>) so snippet can inject ProgressCircle
- Snippet now follows React pattern: import from @seed-design/lynx-react and wrap with ProgressCircle on loading
