+++ docs/public/__registry__/lynx/ui/action-button.json
@@ -1,8 +1,7 @@
 {
   "id": "action-button",
   "dependencies": [
-    "@seed-design/lynx-css",
-    "clsx",
+    "@seed-design/lynx-react",
     "react"
   ],
   "innerDependencies": [
@@ -16,7 +15,7 @@
   "snippets": [
     {
       "path": "action-button.tsx",
-      "content": "/**\n * @file ui:action-button\n * @requires @seed-design/lynx-react@~0.1.0-alpha.0\n * @requires @seed-design/lynx-css@~0.1.0-alpha.0\n **/\n\nimport { actionButton } from \"@seed-design/lynx-css/recipes/action-button\";\nimport type { ActionButtonVariantProps } from \"@seed-design/lynx-css/recipes/action-button\";\nimport clsx from \"clsx\";\nimport * as React from \"react\";\nimport { ProgressCircle } from \"./progress-circle\";\n\nexport interface ActionButtonProps extends Omit<ActionButtonVariantProps, \"layout\"> {\n  children?: React.ReactNode;\n  disabled?: boolean;\n  loading?: boolean;\n  className?: string;\n  flexGrow?: number;\n  bindtap?: () => void;\n  \"main-thread:bindtap\"?: () => void;\n}\n\n/**\n * @see https://seed-design.io/lynx/components/action-button\n */\nexport const ActionButton = React.forwardRef<unknown, ActionButtonProps>((props, ref) => {\n  const [variantProps, restProps] = actionButton.splitVariantProps(props);\n  const {\n    children,\n    className,\n    flexGrow,\n    loading = false,\n    bindtap,\n    \"main-thread:bindtap\": mainThreadBindtap,\n    ...nativeProps\n  } = restProps;\n\n  const { disabled = false } = variantProps;\n  const classes = actionButton({\n    ...variantProps,\n    layout: \"withText\",\n    loading: loading ? true : undefined,\n    disabled: disabled ? true : undefined,\n  });\n  const isInteractive = !disabled && !loading;\n\n  return (\n    <view\n      {...(ref ? { ref: ref as React.Ref<SVGViewElement> } : {})}\n      className={clsx(classes.root, className)}\n      style={flexGrow != null ? { flexGrow } : undefined}\n      {...(isInteractive && bindtap && { bindtap })}\n      {...(isInteractive &&\n        mainThreadBindtap && {\n          \"main-thread:bindtap\": mainThreadBindtap,\n        })}\n      {...nativeProps}\n    >\n      {loading ? (\n        <>\n          <view\n            style={{\n              position: \"absolute\",\n              display: \"flex\",\n              alignItems: \"center\",\n              justifyContent: \"center\",\n              inset: 0,\n            }}\n          >\n            <ProgressCircle size=\"inherit\" tone=\"inherit\" />\n          </view>\n          <text className={classes.text} style={{ opacity: 0 }}>\n            {children}\n          </text>\n        </>\n      ) : (\n        <text className={classes.text}>{children}</text>\n      )}\n    </view>\n  );\n});\nActionButton.displayName = \"ActionButton\";\n\n/**\n * This file is a snippet from SEED Design, helping you get started quickly with @seed-design/* packages.\n * You can extend this snippet however you want.\n */\n",
+      "content": "/**\n * @file ui:action-button\n * @requires @seed-design/lynx-react@~0.1.0-alpha.0\n * @requires @seed-design/lynx-css@~0.1.0-alpha.0\n **/\n\nimport {\n  ActionButton as SeedActionButton,\n  type ActionButtonProps as SeedActionButtonProps,\n} from \"@seed-design/lynx-react\";\nimport * as React from \"react\";\nimport { ProgressCircle } from \"./progress-circle\";\n\nexport interface ActionButtonProps extends SeedActionButtonProps {}\n\n/**\n * @see https://seed-design.io/lynx/components/action-button\n */\nexport const ActionButton = React.forwardRef<\n  React.ElementRef<typeof SeedActionButton>,\n  ActionButtonProps\n>(({ loading = false, children, ...otherProps }, ref) => {\n  return (\n    <SeedActionButton ref={ref} loading={loading} {...otherProps}>\n      {loading ? (\n        <>\n          <view\n            style={{\n              position: \"absolute\",\n              display: \"flex\",\n              alignItems: \"center\",\n              justifyContent: \"center\",\n              inset: 0,\n            }}\n          >\n            <ProgressCircle size=\"inherit\" tone=\"inherit\" />\n          </view>\n          <view style={{ opacity: 0, display: \"inherit\", gap: \"inherit\" }}>{children}</view>\n        </>\n      ) : (\n        children\n      )}\n    </SeedActionButton>\n  );\n});\nActionButton.displayName = \"ActionButton\";\n\n/**\n * This file is a snippet from SEED Design, helping you get started quickly with @seed-design/* packages.\n * You can extend this snippet however you want.\n */\n",
       "dependencies": {
         "@seed-design/lynx-react": "~0.1.0-alpha.0",
         "@seed-design/lynx-css": "~0.1.0-alpha.0"
