+++ docs/registry/lynx/ui/action-button.tsx
@@ -1,55 +1,21 @@
-import { actionButton } from "@seed-design/lynx-css/recipes/action-button";
-import type { ActionButtonVariantProps } from "@seed-design/lynx-css/recipes/action-button";
-import clsx from "clsx";
+import {
+  ActionButton as SeedActionButton,
+  type ActionButtonProps as SeedActionButtonProps,
+} from "@seed-design/lynx-react";
 import * as React from "react";
 import { ProgressCircle } from "./progress-circle";
 
-export interface ActionButtonProps extends Omit<ActionButtonVariantProps, "layout"> {
-  children?: React.ReactNode;
-  disabled?: boolean;
-  loading?: boolean;
-  className?: string;
-  flexGrow?: number;
-  bindtap?: () => void;
-  "main-thread:bindtap"?: () => void;
-}
+export interface ActionButtonProps extends SeedActionButtonProps {}
 
 /**
  * @see https://seed-design.io/lynx/components/action-button
  */
-export const ActionButton = React.forwardRef<unknown, ActionButtonProps>((props, ref) => {
-  const [variantProps, restProps] = actionButton.splitVariantProps(props);
-  const {
-    children,
-    className,
-    flexGrow,
-    loading = false,
-    bindtap,
-    "main-thread:bindtap": mainThreadBindtap,
-    ...nativeProps
-  } = restProps;
-
-  const { disabled = false } = variantProps;
-  const classes = actionButton({
-    ...variantProps,
-    layout: "withText",
-    loading: loading ? true : undefined,
-    disabled: disabled ? true : undefined,
-  });
-  const isInteractive = !disabled && !loading;
-
+export const ActionButton = React.forwardRef<
+  React.ElementRef<typeof SeedActionButton>,
+  ActionButtonProps
+>(({ loading = false, children, ...otherProps }, ref) => {
   return (
-    <view
-      {...(ref ? { ref: ref as React.Ref<SVGViewElement> } : {})}
-      className={clsx(classes.root, className)}
-      style={flexGrow != null ? { flexGrow } : undefined}
-      {...(isInteractive && bindtap && { bindtap })}
-      {...(isInteractive &&
-        mainThreadBindtap && {
-          "main-thread:bindtap": mainThreadBindtap,
-        })}
-      {...nativeProps}
-    >
+    <SeedActionButton ref={ref} loading={loading} {...otherProps}>
       {loading ? (
         <>
           <view
@@ -63,14 +29,12 @@ export const ActionButton = React.forwardRef<unknown, ActionButtonProps>((props,
           >
             <ProgressCircle size="inherit" tone="inherit" />
           </view>
-          <text className={classes.text} style={{ opacity: 0 }}>
-            {children}
-          </text>
+          <view style={{ opacity: 0, display: "inherit", gap: "inherit" }}>{children}</view>
         </>
       ) : (
-        <text className={classes.text}>{children}</text>
+        children
       )}
-    </view>
+    </SeedActionButton>
   );
 });
 ActionButton.displayName = "ActionButton";
