+++ packages/lynx-react/src/components/ActionButton/ActionButton.tsx
@@ -7,7 +7,6 @@ import * as React from "react";
  * @platform Lynx
  *
  * 미지원 기능 (Lynx 3.7 SVG 지원 후 추가 예정):
- * - loading: SVG 기반 spinner 필요
  * - layout: "iconOnly": SVG 아이콘 렌더링 필요
  * - PrefixIcon / SuffixIcon: SVG 아이콘 렌더링 필요
  *
@@ -15,9 +14,10 @@ import * as React from "react";
  * - color / fontWeight props: CSS variable 동적 주입 제한
  * - bleedX / bleedY props: CSS variable 동적 주입 제한
  */
-export interface ActionButtonProps extends Omit<ActionButtonVariantProps, "loading" | "layout"> {
+export interface ActionButtonProps extends Omit<ActionButtonVariantProps, "layout"> {
   children?: React.ReactNode;
   disabled?: boolean;
+  loading?: boolean;
   className?: string;
   flexGrow?: number;
   bindtap?: () => void;
@@ -30,6 +30,7 @@ export const ActionButton = React.forwardRef<unknown, ActionButtonProps>((props,
     children,
     className,
     flexGrow,
+    loading = false,
     bindtap,
     "main-thread:bindtap": mainThreadBindtap,
     ...nativeProps
@@ -39,9 +40,10 @@ export const ActionButton = React.forwardRef<unknown, ActionButtonProps>((props,
   const classes = actionButton({
     ...variantProps,
     layout: "withText",
+    loading: loading ? true : undefined,
     disabled: disabled ? true : undefined,
   });
-  const isInteractive = !disabled;
+  const isInteractive = !disabled && !loading;
 
   return (
     <view
@@ -55,7 +57,7 @@ export const ActionButton = React.forwardRef<unknown, ActionButtonProps>((props,
         })}
       {...nativeProps}
     >
-      <text className={classes.text}>{children}</text>
+      {loading ? children : <text className={classes.text}>{children}</text>}
     </view>
   );
 });
